fix: ffmpeg error is now actually human readable
When ffmpeg/ffprobe fails, the package we're using likes to send the
error to us with a bunch of additional useless information (compilation
flags???). With this commit, we now search the output string for an
error message (json encoded string), and if we find it we use that to
populate the message of our trouble. This results in a significantly
more readable error message (and much.. much shorter too)