fix: deadlock from workers refusing to wake
Not really sure what causes it, however it's clear that an easy fix is
to send the wakeup signal in a non-blocking select statement. However
WHY the worker is reporting it's status as 'Sleeping', but is not
actually listening to the channel for wakeup signals is strange. Could
be a quirk of it being in a different thread where the instructions are
slightly out of sync.