feat: imported items now checked for changes
Currently, if you were to copy a file in to the import directory, this
system could pick it up half-way through and begin formatting an
incomplete file. This change means the files 'last modified' time is
checked frequently and is only advanced from 'Import' to 'Title' stage
once it's been unmodified for 2 minutes or more.